Update djthorn-trading-indicators.pinescript
Added a rule when a vector candle should be red or green. 
Now it must first meet the volume rule of being more then 250% of the average volume of the previous 50 candles.
A green candle must have its open be equal to that of the low
A red candle must have its open be equal to its high.